Move card related classes into its own package
