+++ src/main/java/no/uib/inf112/core/ui/cards/CardActor.java
@@ -1,9 +1,10 @@
-package no.uib.inf112.core.ui;
+package no.uib.inf112.core.ui.cards;
 
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageTextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import no.uib.inf112.core.player.Card;
+import no.uib.inf112.core.ui.UIHandler;
 import org.jetbrains.annotations.Nullable;
 
 /**
