+++ src/main/java/no/uib/inf112/core/ui/cards/CardActorSlot.java
@@ -1,6 +1,7 @@
-package no.uib.inf112.core.ui;
+package no.uib.inf112.core.ui.cards;
 
 import com.badlogic.gdx.scenes.scene2d.utils.DragAndDrop;
+import no.uib.inf112.core.ui.CardContainer;
 import org.jetbrains.annotations.NotNull;
 
 /**
